requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.39.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
